Add core module for core functionality
