+++ lib/shopify-cli/core.rb
@@ -1,0 +1,10 @@
+module ShopifyCli
+  module Core
+    autoload :EntryPoint, 'shopify-cli/core/entry_point'
+    autoload :Executor, 'shopify-cli/core/executor'
+    autoload :Finalize, 'shopify-cli/core/finalize'
+    autoload :HelpResolver, 'shopify-cli/core/help_resolver'
+    autoload :Monorail, 'shopify-cli/core/monorail'
+    autoload :Update, 'shopify-cli/core/update'
+  end
+end
