+++ test/shopify-cli/commands/load_system_test.rb
@@ -4,7 +4,7 @@ module ShopifyCli
   module Commands
     class LoadSystemTest < MiniTest::Test
       def test_loads_system
-        ShopifyCli::Finalize.expects(:reload_shopify_from).with(ShopifyCli::INSTALL_DIR)
+        ShopifyCli::Core::Finalize.expects(:reload_shopify_from).with(ShopifyCli::INSTALL_DIR)
         capture_io do
           run_cmd('load-system')
         end
