+++ test/shopify-cli/commands/update_test.rb
@@ -4,7 +4,7 @@ module ShopifyCli
   module Commands
     class UpdateTest < MiniTest::Test
       def test_calls_update
-        ShopifyCli::Update.expects(:check_now).with(
+        ShopifyCli::Core::Update.expects(:check_now).with(
           restart_command_after_update: false,
           ctx: @context,
         )
