+++ test/shopify-cli/smoke_test.rb
@@ -5,7 +5,7 @@ module ShopifyCli
     def test_exit_non_zero
       assert_nothing_raised do
         capture_io do
-          ShopifyCli::EntryPoint.call(['help'])
+          ShopifyCli::Core::EntryPoint.call(['help'])
         end
       end
     end
